restructure atom and system to separate modules
